zuiku/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
psycopg2-binary>=2.9.11
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
httpx==0.25.0